Move serializers/deserializers into json package
